+++ UI-Swift-Sample/UI-Swift-Sample/AppDelegate.swift
@@ -7,7 +7,7 @@
 //
 
 import UIKit
-import NearITSDK
+import NearITSDKSwift
 import OHHTTPStubs
 import NearUIBinding
 import UserNotifications
@@ -53,22 +53,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         NITManager.default().stop()
     }
 
-    func applicationDidEnterBackground(_ application: UIApplication) {
-        // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
-        // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
-    }
-
-    func applicationWillEnterForeground(_ application: UIApplication) {
-        // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
-    }
-
     func applicationDidBecomeActive(_ application: UIApplication) {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
         NITManager.default().start()
-    }
-
-    func applicationWillTerminate(_ application: UIApplication) {
-        // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
     }
 
     func enableStubForHistory(_ enabled: Bool) {
@@ -115,16 +102,20 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
     
     @available(iOS 10.0, *)
     func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-        completionHandler(.alert)
+        NearManager.shared.userNotificationCenter(center, willPresent: notification, withCompletionHandler: completionHandler)
     }
     
   @available(iOS 10.0, *)
   func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
-    let userInfo = response.notification.request.content.userInfo
-    if let ui = userInfo as? [String : Any] {
-      NITManager.default().processRecipe(userInfo: ui, completion: { (content, trackingInfo, error) in
-        self.handleNearContent(content: content!, trackingInfo: trackingInfo!)
-      })
+        NearManager.shared.showContentFrom(response) { (content, trackingInfo, error) in
+            if error != nil {
+                // there was an error
+            }
+            if customJson = content as? NITCustomJSON {
+                // handle the custom JSON
+            }
+        }
+        completionHandler()
     }
   }
 }